完善创建队伍业务逻辑 添加队伍状态枚举类 为update/insert/delete操作添加事务管理 存储建表语句 修改各种小问题。。。

+++ src/main/java/com/pikachu/usercenter/service/TeamService.java
@@ -21,7 +21,7 @@ public interface TeamService extends IService<Team> {
 
     TeamVO updateTeam(TeamUpdateRequest teamUpdateRequest, HttpServletRequest request);
 
-    TeamVO getTeamById(Integer teamId);
+    TeamVO getTeamById(Long teamId);
 
-    IPage<TeamVO> searchTeams(Integer current, Integer size);
+    IPage<TeamVO> searchTeams(Long current, Long size);
 }
